@@ -1,16 +1,18 @@
-import React from "react";
-import ReactDOM from "react-dom";
+import React from 'react'
+import ReactDOM from 'react-dom/client'
 import { BrowserRouter } from "react-router-dom";
 import "./styles/globals.css";
 import App from "./App";
 import { Provider } from "react-redux";
 import store from "./Redux/store";
 
-ReactDOM.render(
+
+ReactDOM.createRoot(document.getElementById('root')).render(
   <Provider store={store}>
     <BrowserRouter>
       <App />
     </BrowserRouter>
   </Provider>,
-  document.getElementById("root")
-);
+)
+
+
